+++ src/environments/static_environments.py
@@ -1,16 +1,17 @@
-import numpy as np
+# import numpy as np
 
 from types import SimpleNamespace
 
 import src.generating_functions as GeneratingFunctions
 import src.noise as Noise 
 
-from src.maths import quadratic
+# from src.maths import quadratic
 
 
-class StaticLinearEnvironment:
-    def __init__(self, params: dict) -> None:
+class StaticEnvironment:
+    def __init__(self, params: dict, nonlinear: bool = False) -> None:
         self.params = SimpleNamespace(**params)
+        self.nonlinear = nonlinear
         
     def build(self, x_star: float):
         
@@ -21,10 +22,16 @@ class StaticLinearEnvironment:
         self.noise = Noise.zero_centered_normal(scale=self.params.y_star_std)
         
         # Define generating function
-        self.generating_function = GeneratingFunctions.linear(
-            x_star=x_star, 
-            intercept=self.params.beta_0_star, 
-            slope=self.params.beta_1_star)
+        if self.nonlinear:
+            self.generating_function = GeneratingFunctions.nonlinear_quadratic(
+                x_star=x_star, 
+                intercept=self.params.beta_0_star, 
+                slope=self.params.beta_1_star)
+        else:
+            self.generating_function = GeneratingFunctions.linear(
+                x_star=x_star, 
+                intercept=self.params.beta_0_star, 
+                slope=self.params.beta_1_star)
         
     def store_history(self):
         keys = ["x_star", "noise", "generating_function"]
@@ -33,19 +40,19 @@ class StaticLinearEnvironment:
     def generate(self):
         return self.generating_function + self.noise
 
-class StaticNonlinearEnvironment:
-    def __init__(self, params: dict) -> None:
-        self.params = SimpleNamespace(**params)
+# class StaticNonlinearEnvironment:
+#     def __init__(self, params: dict) -> None:
+#         self.params = SimpleNamespace(**params)
         
-    def noise(self) -> float:
-        return np.random.normal(loc=0, scale=self.params.y_star_std)
+#     def noise(self) -> float:
+#         return np.random.normal(loc=0, scale=self.params.y_star_std)
     
-    def phi(self, x_star: float) -> float:
-        return x_star**2
+#     def phi(self, x_star: float) -> float:
+#         return x_star**2
     
-    def generating_function(self, x_star: float) -> float:
-        return self.params.beta_1_star * quadratic(x_star) + self.params.beta_0_star
+#     def generating_function(self, x_star: float) -> float:
+#         return self.params.beta_1_star * quadratic(x_star) + self.params.beta_0_star
     
-    def generate(self, x_star: float) -> float:
-        return self.generating_function(x_star) + self.noise()
+#     def generate(self, x_star: float) -> float:
+#         return self.generating_function(x_star) + self.noise()
